third_party/WebKit/LayoutTests/plugins/focus-change-1-no-change.html

<script src="../resources/js-test.js"></script>
<p>Test no change of focus inside plugin onfocus and onblur events</p>
<input id="input1" type="text" onfocus="i1Focus()" onblur="i1Blur()"></input>
<input id="input2" type="text" onfocus="i2Focus()" onblur="i2Blur()"></input>
<embed id="plugin" type="application/x-blink-test-plugin" onfocus="pFocus()" onblur="pBlur()"></embed>
<script>
function i1Focus() { debug('i1focus'); }
function i1Blur() { debug('i1blur'); }
function i2Focus() { debug('i2focus'); }
function i2Blur() { debug('i2blur'); }
function pFocus() { debug('pfocus'); }
function pBlur() { debug('pblur'); }

if (testRunner)
  testRunner.waitUntilDone();

document.getElementById('input1').focus();
document.getElementById('plugin').focus();
document.getElementById('input2').focus();

setTimeout(
    function() {
      if (testRunner)
        testRunner.notifyDone();
    },
    0);
</script>
