third_party/blink/web_tests/platform/mac/plugins/focus-change-1-no-change-expected.txt

CONSOLE MESSAGE: Blink Test Plugin: initializing
CONSOLE MESSAGE: Blink Test Plugin: DidChangeFocus(true)
i1focus
i1blur
pfocus
pblur
i2focus
PASS successfullyParsed is true

TEST COMPLETE


Test no change of focus inside plugin onfocus and onblur events

  
